requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiosqlite>=0.19.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
